chore: KEEP-579 move protocol-coverage tests from integration to e2e
These tests build a workflow per protocol action, POST it to the keeperhub
server's webhook endpoint at PROTOCOL_E2E_BASE_URL, and poll
workflow_executions for success. They are not integration tests by the
project's definition - they need a running keeperhub server, a connected
DB, and a Turnkey-provisioned wallet. The integration job has none of
those, so the suite has been silently skipping in CI since it landed
(KEEP-458 protocol coverage).

Move it where it belongs:

  tests/integration/protocol-coverage/aave-v3/sepolia/coverage.test.ts
  tests/integration/protocol-coverage/superfluid/sepolia/coverage.test.ts
  tests/integration/protocol-coverage/_shared/{funding,run-fixture,setup}.ts

  ->

  tests/e2e/vitest/protocol-coverage/aave-v3/sepolia/coverage.test.ts
  tests/e2e/vitest/protocol-coverage/superfluid/sepolia/coverage.test.ts
  tests/e2e/vitest/protocol-coverage/_shared/{funding,run-fixture,setup}.ts

git mv preserves history.

Gating: switch from describe.skipIf(!SEPOLIA_RPC_URL) to the convention
already used by sibling e2e tests (full-pipeline.test.ts,
gas-strategy.test.ts):

  const SKIP_INFRA_TESTS =
    !process.env.DATABASE_URL || process.env.SKIP_INFRA_TESTS === 'true';

This is more honest - the test cannot do anything without DB access
(needs to insert workflows + read execution rows), and the existing
SKIP_INFRA_TESTS env var is the project's standard escape hatch.
The previous SEPOLIA_RPC_URL gate was a proxy for 'is the environment
up at all'; the test code itself never reads that variable - the server
the test calls into reads CHAIN_RPC_CONFIG instead.

Discovery: pnpm test:e2e:vitest globs tests/e2e/vitest/**, so the moved
files are picked up automatically. Confirmed locally - the runner now
sees 16 files (was 14). Also updated the test:protocol convenience
script in package.json to point at the new path.

Import updates: the two unit tests that import the helpers
(tests/unit/{faucet-args,plan-phase-fixtures}.test.ts) and one doc
comment in lib/test-data/chain-test-data.ts.

Caveat for reviewer: the e2e-vitest-ephemeral CI job is currently
DISABLED pending KEEP-1351 (auth/rate-limit infrastructure stabilisation).
This move puts the coverage tests in the right place architecturally,
but they will only run on CI once that job is re-enabled. The previous
state (skipping in test-integration) gave the same zero-coverage outcome;
the difference is that the tests are now correctly categorised and will
join the e2e run as soon as the job comes back.
