+++ tests/e2e/vitest/protocol-coverage/_shared/run-fixture.ts
