+++ tests/e2e/vitest/protocol-coverage/_shared/setup.ts
